backend/requirements.txt
fastapi
uvicorn
google-genai
python-multipart
pillow
pydub
speechrecognition
pymongo
python-dotenv
requests
